update: fix auth and user api

+++ src/services/user.service.ts
@@ -1,47 +1,60 @@
 import { HTTPException } from "hono/http-exception";
 import prisma from "../lib/prisma";
-import redis from "../lib/redis";
-import { UserRequest, UserUpdateRequest } from "../zod/user.zod";
+import { UserUpdateRequest } from "../zod/user.zod";
 import bcrypt from "bcrypt";
 
-export function createUser({ name, email, password }: UserRequest) {
-  return prisma.user.create({
-    data: {
-      name,
-      email,
-      password: bcrypt.hashSync(password, 10),
-    },
-  });
-}
-
-export function getUsers() {
+export async function getUsers() {
   return prisma.user.findMany();
 }
 
-export function getUser({ id }: { id: string }) {
+export async function getUser({ id }: { id: string }) {
   return prisma.user.findFirst({
     where: { id },
+    include: { address: true },
   });
 }
 
-export function updateUser({
+export async function updateUser({
   name,
   password,
   id,
+  address,
 }: UserUpdateRequest & { id: string }) {
   return prisma.user.update({
     data: {
       name,
       password: password ? bcrypt.hashSync(password, 10) : undefined,
+      address: address
+        ? {
+            upsert: {
+              create: address,
+              update: address,
+            },
+          }
+        : undefined,
     },
     where: { id },
+    include: { address: true },
   });
 }
 
 export async function deleteUser({ id }: { id: string }) {
+  const user = await prisma.user.findFirst({
+    where: { id },
+    include: { address: true },
+  });
+
+  if (!user) {
+    throw new HTTPException(404, { message: "User not found" });
+  }
+
+  if (user.address) {
+    await prisma.address.delete({
+      where: { userId: id },
+    });
+  }
+
   await prisma.user.delete({
-    where: {
-      id,
-    },
+    where: { id },
   });
 }
